WorkspaceColumnsにopen stateを定義 #37
