+++ client/src/components/Column/WorkspaceColumns.js
@@ -1,8 +1,10 @@
-import React, { useEffect } from "react";
+import React, { useState, useEffect } from "react";
 import AddColumn from "../../containers/Column/AddColumn";
 import ColumnList from "./ColumnList";
 
 const WorkspaceColumns = ({ columns, getColumns, workspace }) => {
+  const [open, setOpen] = useState(false);
+
   useEffect(() => {
     getColumns(workspace.id);
   }, [getColumns, workspace.id]);
@@ -11,10 +13,25 @@ const WorkspaceColumns = ({ columns, getColumns, workspace }) => {
     return workspace.columns.length > 0;
   };
 
+  const handleClickOpen = () => {
+    setOpen(true);
+  };
+
+  const handleClose = () => {
+    setOpen(false);
+  };
+
   if (columnsExist()) {
     return <ColumnList columns={columns} />;
   } else {
-    return <AddColumn workspace={workspace} />;
+    return (
+      <AddColumn
+        workspace={workspace}
+        open={open}
+        handleClose={handleClose}
+        handleClickOpen={handleClickOpen}
+      />
+    );
   }
 };
